fix customer_id to BigInteger

+++ app/database/models/tables/admin.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Index, Integer, Column, Float, text
+from sqlalchemy import Index, Integer, Column, Float, text, BigInteger
 
 from database.engine import Base
 
@@ -7,7 +7,7 @@ class Admin(Base):
     __tablename__ = 'admins'
 
     admin_id: int = Column(Integer, primary_key=True, autoincrement=True)
-    chat_id: int = Column(Integer, nullable=False)
+    chat_id: int = Column(BigInteger, nullable=False)
     debt: float = Column(
         Float(precision=7, decimal_return_scale=2),
         nullable=False,
