+++ app/database/models/tables/sub_admin.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Index
+from sqlalchemy import Column, Integer, ForeignKey, Index, BigInteger
 from sqlalchemy.orm import relationship
 
 from database.engine import Base
@@ -8,7 +8,7 @@ class SubAdmin(Base):
     __tablename__ = 'sub_admins'
 
     id: int = Column(Integer, primary_key=True, autoincrement=True)
-    chat_id: int = Column(Integer, nullable=False)
+    chat_id: int = Column(BigInteger, nullable=False)
 
     tg_bot_id: int = Column(Integer, ForeignKey('tg_bots.tg_bot_id'))
     tg_bot = relationship('TGBot', backref='sub_admins')
