+++ app/database/models/tables/tg_bot.py
@@ -12,7 +12,7 @@ class TGBot(Base):
     tg_bot_id: int = Column(BigInteger, primary_key=True, autoincrement=False)
     tg_token: int = Column(String(46), nullable=False)
     start_message: Optional[str] = Column(Text, nullable=True)
-    qiwi_txn: Optional[int] = Column(Integer, nullable=True)
+    qiwi_txn: Optional[int] = Column(BigInteger, nullable=True)
 
     admin_id: int = Column(
         Integer,
